requirement.txt
absl-py==2.0.0
accelerate==0.24.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
async-lru==2.0.4
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
chex==0.1.84
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
debugpy==1.8.0
decorator==5.1.1
diffusers @ file:///C:/Users/user/diffusers
distlib==0.3.7
easydict==1.10
etils==1.5.2
executing==2.0.1
filelock==3.13.1
Flask==3.0.0
flax==0.7.5
fonttools==4.43.1
fqdn==1.5.1
fsspec==2023.10.0
ftfy==6.1.1
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.1
insightface==0.7.3
ipykernel==6.26.0
ipython==8.17.2
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.20
jaxlib==0.4.20
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jupyter-events==0.9.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab_server==2.25.0
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
omegaconf==2.3.0
onnx==1.14.1
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.2
overrides==7.4.0
packaging==23.2
parso==0.8.3
Pillow==10.1.0
platformdirs==3.11.0
prettytable==3.9.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
Send2Trash==1.8.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorstore==0.1.47
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==2.0.7
virtualenv==20.24.6
wcwidth==0.2.8
webcolors==1.13
websocket-client==1.6.4
Werkzeug==3.0.1
wfastcgi==3.0.0
wget==3.2
xformers==0.0.22.post7
zipp==3.17.0
